# docs/api/openapi-generated.yaml
components:
  schemas:
...
    CourseExamsPage:
      properties:
        course_id:
          format: uuid
          title: Course Id
          type: string
        course_name:
          title: Course Name
          type: string
        exams:
          items:
            $ref: '#/components/schemas/ExamInList'
          title: Exams
          type: array
      required:
      - course_id
      - course_name
      - exams
      title: CourseExamsPage
      type: object
...
    Exam:
      properties:
        duration_minutes:
          description: Duration of the exam in minutes
          example: 60
          title: Duration Minutes
          type: integer
        end_at:
          example: '2027-10-08T10:00:00Z'
          format: date-time
          title: End At
          type: string
        id:
          example: a1b2c3d4-e5f6-7890-1234-567890abcdef
          format: uuid
          title: Id
          type: string
        instructions:
          anyOf:
          - type: string
          - type: 'null'
          example: Іспит складається з 20 теоретичних питань.
          title: Instructions
        last_attempt_id:
          anyOf:
          - type: string
          - type: 'null'
          description: ID of the last attempt for this exam by the current user
          example: c7a1c7e2-4a2c-4b6e-8e7f-9d3c5f2b1a8e
          title: Last Attempt Id
        max_attempts:
          example: 3
          title: Max Attempts
          type: integer
        owner_id:
          example: c7a1c7e2-4a2c-4b6e-8e7f-9d3c5f2b1a8e
          format: uuid
          title: Owner Id
          type: string
        pass_threshold:
          example: 75
          title: Pass Threshold
          type: integer
        published:
          default: false
          description: Whether exam is published
          title: Published
          type: boolean
        question_count:
          default: 0
          description: Number of available questions
          example: 20
          title: Question Count
          type: integer
        start_at:
          example: '2024-10-08T10:00:00Z'
          format: date-time
          title: Start At
          type: string
        status:
          description: 'Exam status: draft, published, open, or closed'
          example: draft
          title: Status
          type: string
        title:
          example: Вступ до Docker
          title: Title
          type: string
        user_attempts_count:
          anyOf:
          - type: integer
          - type: 'null'
          description: Number of attempts made by the current user for this exam
          example: 2
          title: User Attempts Count
      required:
      - id
      - title
      - start_at
      - end_at
      - duration_minutes
      - max_attempts
      - pass_threshold
      - owner_id
      - status
      title: Exam
      type: object
...
    ExamCreate:
      properties:
        duration_minutes:
          default: 60
          description: Duration of the exam in minutes
          example: 120
          title: Duration Minutes
          type: integer
        end_at:
          description: End datetime (UTC)
          example: '2027-10-08T10:00:00Z'
          format: date-time
          title: End At
          type: string
        instructions:
          anyOf:
          - maxLength: 2000
            type: string
          - type: 'null'
          description: Markdown/HTML instructions
          example: Іспит складається з 20 теоретичних питань.
          title: Instructions
        max_attempts:
          default: 1
          description: Max attempts per user
          example: 3
          maximum: 10.0
          minimum: 1.0
          title: Max Attempts
          type: integer
        owner_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          description: Instructor user id (automatically set from token if not provided)
          example: c7a1c7e2-4a2c-4b6e-8e7f-9d3c5f2b1a8e
          title: Owner Id
        pass_threshold:
          default: 60
          description: Passing threshold in percent
          example: 75
          maximum: 100.0
          minimum: 0.0
          title: Pass Threshold
          type: integer
        start_at:
          description: Start datetime (UTC)
          example: '2026-10-08T10:00:00Z'
          format: date-time
          title: Start At
          type: string
        title:
          description: Exam title
          example: Вступ до Docker
          maxLength: 100
          minLength: 3
          title: Title
          type: string
      required:
      - title
      - start_at
      - end_at
      title: ExamCreate
      type: object
    ExamInList:
      properties:
        average_grade:
          anyOf:
          - type: number
          - type: 'null'
          title: Average Grade
        id:
          format: uuid
          title: Id
          type: string
        pending_reviews:
          default: 0
          title: Pending Reviews
          type: integer
        questions_count:
          default: 0
          title: Questions Count
          type: integer
        status:
          $ref: '#/components/schemas/ExamStatusEnum'
        students_completed:
          title: Students Completed
          type: string
        title:
          title: Title
          type: string
      required:
      - id
      - title
      - status
      - students_completed
      title: ExamInList
      type: object
...
    ExamParticipantAttendanceUpdate:
      properties:
        status:
          description: Статус присутності студента
          enum:
          - present
          - absent
          title: Status
          type: string
      required:
      - status
      title: ExamParticipantAttendanceUpdate
      type: object
    ExamParticipantCreate:
      properties:
        course_id:
          description: ID курсу для перевірки зарахування
          format: uuid
          title: Course Id
          type: string
        user_id:
          description: ID студента, якого додаємо до іспиту
          format: uuid
          title: User Id
          type: string
      required:
      - user_id
      - course_id
      title: ExamParticipantCreate
      type: object
    ExamParticipantResponse:
      properties:
        attendance_status:
          enum:
          - unknown
          - present
          - absent
          title: Attendance Status
          type: string
        exam_id:
          format: uuid
          title: Exam Id
          type: string
        is_active:
          title: Is Active
          type: boolean
        joined_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Joined At
        removed_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Removed At
        user_id:
          format: uuid
          title: User Id
          type: string
      required:
      - exam_id
      - user_id
      - is_active
      - attendance_status
      title: ExamParticipantResponse
      type: object
...
    ExamStatusEnum:
      enum:
      - draft
      - published
      - open
      - closed
      title: ExamStatusEnum
      type: string
    ExamUpdate:
      properties:
        duration_minutes:
          anyOf:
          - type: integer
          - type: 'null'
          description: Duration of the exam in minutes
          example: 120
          title: Duration Minutes
        end_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          description: End datetime (UTC)
          example: '2028-02-20T18:00:00Z'
          title: End At
        instructions:
          anyOf:
          - maxLength: 2000
            type: string
          - type: 'null'
          description: Markdown/HTML instructions
          example: 'Оновлені інструкції: додано практичне завдання.'
          title: Instructions
        max_attempts:
          anyOf:
          - maximum: 10.0
            minimum: 1.0
            type: integer
          - type: 'null'
          description: Max attempts per user
          example: 2
          title: Max Attempts
        pass_threshold:
          anyOf:
          - maximum: 100.0
            minimum: 0.0
            type: integer
          - type: 'null'
          description: Passing threshold in percent
          example: 80
          title: Pass Threshold
        published:
          anyOf:
          - type: boolean
          - type: 'null'
          description: Publish exam (true/false)
          title: Published
        start_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          description: Start datetime (UTC)
          example: '2026-11-15T09:00:00Z'
          title: Start At
        title:
          anyOf:
          - maxLength: 100
            minLength: 3
            type: string
          - type: 'null'
          description: Exam title
          example: 'Docker: Просунутий рівень'
          title: Title
      title: ExamUpdate
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    LoginRequest:
      description: Схема для запиту на вхід у систему.
      properties:
        email:
          format: email
          title: Email
          type: string
        password:
          title: Password
          type: string
      required:
      - email
      - password
      title: LoginRequest
      type: object
    LoginResponse:
      description: Повна відповідь, що надсилається клієнту після успішного входу.
      properties:
        access_token:
          title: Access Token
          type: string
        token_type:
          default: bearer
          title: Token Type
          type: string
        user:
          $ref: '#/components/schemas/UserResponse'
      required:
      - access_token
      - user
      title: LoginResponse
      type: object
...
    RegisterRequest:
      description: Схема для запиту на реєстрацію нового користувача.
      properties:
        email:
          format: email
          title: Email
          type: string
        first_name:
          title: First Name
          type: string
        last_name:
          title: Last Name
          type: string
        major_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Major Id
        password:
          title: Password
          type: string
        patronymic:
          anyOf:
          - type: string
          - type: 'null'
          title: Patronymic
      required:
      - email
      - password
      - first_name
      - last_name
      title: RegisterRequest
      type: object
...
    UserResponse:
      description: Схема для відображення публічної інформації про користувача.
      properties:
        avatar_url:
          anyOf:
          - format: uri
            maxLength: 2083
            minLength: 1
            type: string
          - type: 'null'
          title: Avatar Url
        email:
          format: email
          title: Email
          type: string
        full_name:
          title: Full Name
          type: string
        id:
          format: uuid
          title: Id
          type: string
        roles:
          default: []
          items:
            type: string
          title: Roles
          type: array
        user_major:
          anyOf:
          - type: string
          - type: 'null'
          title: User Major
      required:
      - id
      - email
      - full_name
      title: UserResponse
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  contact:
    email: team@example.com
    name: Team
  description: Code-first FastAPI spec for an online examination platform.
  license:
    name: MIT
  title: Online Exams API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /api/auth/login:
    post:
      operationId: login_api_auth_login_post
      parameters:
      - description: API version (e.g., 1.0)
        in: query
        name: api-version
        required: false
        schema:
          default: '1.0'
          description: API version (e.g., 1.0)
          title: Api-Version
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: User login
      tags:
      - Auth
  /api/auth/register:
    post:
      description: Реєстрація нового користувача. За замовчуванням призначається роль
        'student'.
      operationId: register_api_auth_register_post
      parameters:
      - description: API version (e.g., 1.0)
        in: query
        name: api-version
        required: false
        schema:
          default: '1.0'
          description: API version (e.g., 1.0)
          title: Api-Version
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: User registration
      tags:
      - Auth
...
  /api/courses/{course_id}/exams:
    get:
      description: Отримує список іспитів для курсу. Доступно для викладачів та наглядачів.
      operationId: list_course_exams_api_courses__course_id__exams_get
      parameters:
      - in: path
        name: course_id
        required: true
        schema:
          format: uuid
          title: Course Id
          type: string
      - description: API version (e.g., 1.0)
        in: query
        name: api-version
        required: false
        schema:
          default: '1.0'
          description: API version (e.g., 1.0)
          title: Api-Version
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseExamsPage'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Список іспитів для курсу (для викладача та наглядача)
      tags:
      - Courses
...
  /api/exams:
    get:
      operationId: list_course_exams_api_exams_get
      parameters:
      - description: ID курсу для фільтрації іспитів
        in: query
        name: course_id
        required: true
        schema:
          description: ID курсу для фільтрації іспитів
          format: uuid
          title: Course Id
          type: string
      - description: API version (e.g., 1.0)
        in: query
        name: api-version
        required: false
        schema:
          default: '1.0'
          description: API version (e.g., 1.0)
          title: Api-Version
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseExamsPage'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Список іспитів для курсу (лише для викладача)
      tags:
      - Exams
    post:
      operationId: create_exam_api_exams_post
      parameters:
      - description: API version (e.g., 1.0)
        in: query
        name: api-version
        required: false
        schema:
          default: '1.0'
          description: API version (e.g., 1.0)
          title: Api-Version
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExamCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exam'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Create exam
      tags:
      - Exams
...
  /api/exams/{exam_id}:
    delete:
      operationId: delete_exam_api_exams__exam_id__delete
      parameters:
      - in: path
        name: exam_id
        required: true
        schema:
          format: uuid
          title: Exam Id
          type: string
      - description: API version (e.g., 1.0)
        in: query
        name: api-version
        required: false
        schema:
          default: '1.0'
          description: API version (e.g., 1.0)
          title: Api-Version
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Delete exam
      tags:
      - Exams
    get:
      operationId: get_exam_api_exams__exam_id__get
      parameters:
      - in: path
        name: exam_id
        required: true
        schema:
          format: uuid
          title: Exam Id
          type: string
      - description: API version (e.g., 1.0)
        in: query
        name: api-version
        required: false
        schema:
          default: '1.0'
          description: API version (e.g., 1.0)
          title: Api-Version
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exam'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Get exam by id
      tags:
      - Exams
    patch:
      operationId: update_exam_api_exams__exam_id__patch
      parameters:
      - in: path
        name: exam_id
        required: true
        schema:
          format: uuid
          title: Exam Id
          type: string
      - description: API version (e.g., 1.0)
        in: query
        name: api-version
        required: false
        schema:
          default: '1.0'
          description: API version (e.g., 1.0)
          title: Api-Version
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExamUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exam'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Update exam (partial)
      tags:
      - Exams
...
  /api/exams/{exam_id}/participants:
    get:
      operationId: list_participants_api_exams__exam_id__participants_get
      parameters:
      - in: path
        name: exam_id
        required: true
        schema:
          format: uuid
          title: Exam Id
          type: string
      - description: API version (e.g., 1.0)
        in: query
        name: api-version
        required: false
        schema:
          default: '1.0'
          description: API version (e.g., 1.0)
          title: Api-Version
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ExamParticipantResponse'
                title: Response List Participants Api Exams  Exam Id  Participants
                  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Список активних учасників іспиту
      tags:
      - Exam Participants
    post:
      operationId: add_participant_api_exams__exam_id__participants_post
      parameters:
      - in: path
        name: exam_id
        required: true
        schema:
          format: uuid
          title: Exam Id
          type: string
      - description: API version (e.g., 1.0)
        in: query
        name: api-version
        required: false
        schema:
          default: '1.0'
          description: API version (e.g., 1.0)
          title: Api-Version
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExamParticipantCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamParticipantResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Додати студента до іспиту (заборонено, якщо не зарахований на курс)
      tags:
      - Exam Participants
  /api/exams/{exam_id}/participants/{user_id}:
    delete:
      operationId: remove_participant_api_exams__exam_id__participants__user_id__delete
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          format: uuid
          title: User Id
          type: string
      - in: path
        name: exam_id
        required: true
        schema:
          format: uuid
          title: Exam Id
          type: string
      - description: API version (e.g., 1.0)
        in: query
        name: api-version
        required: false
        schema:
          default: '1.0'
          description: API version (e.g., 1.0)
          title: Api-Version
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Видалити студента (завершити активну спробу, якщо йде тест)
      tags:
      - Exam Participants
  /api/exams/{exam_id}/participants/{user_id}/attendance:
    patch:
      operationId: set_attendance_api_exams__exam_id__participants__user_id__attendance_patch
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          format: uuid
          title: User Id
          type: string
      - in: path
        name: exam_id
        required: true
        schema:
          format: uuid
          title: Exam Id
          type: string
      - description: API version (e.g., 1.0)
        in: query
        name: api-version
        required: false
        schema:
          default: '1.0'
          description: API version (e.g., 1.0)
          title: Api-Version
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExamParticipantAttendanceUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamParticipantResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Позначити студента як присутнього або відсутнього
      tags:
      - Exam Participants
...