# src/api/controllers/exam_participants_controller.py
from __future__ import annotations
from fastapi import APIRouter, Depends, status, Path
from sqlalchemy.orm import Session
from uuid import UUID

from src.api.database import get_db
from src.api.schemas.exam_participants import (
    ExamParticipantCreate,
    ExamParticipantResponse,
    ExamParticipantAttendanceUpdate,
)
from src.api.services.exam_participants_service import ExamParticipantsService
from src.utils.auth import require_role
from src.models.users import User
from .versioning import require_api_version

SUPERVISOR_ACCESS_DENIED = "Доступ дозволений лише наглядачам"

class ExamParticipantsController:
    def __init__(self, service: ExamParticipantsService):
        self.service = service
        self.router = APIRouter(
            prefix="/exams/{exam_id}/participants",
            tags=["Exam Participants"],
            dependencies=[Depends(require_api_version)]
        )

        @self.router.get(
            "",
            response_model=list[ExamParticipantResponse],
            summary="Список активних учасників іспиту"
        )
        async def list_participants(
            exam_id: UUID = Path(...),
            db: Session = Depends(get_db),
            current_user: User = Depends(require_role('supervisor')),
        ):
            return self.service.list(db, exam_id)

        @self.router.post(
            "",
            response_model=ExamParticipantResponse,
            status_code=status.HTTP_201_CREATED,
            summary="Додати студента до іспиту (заборонено, якщо не зарахований на курс)"
        )
        async def add_participant(
            payload: ExamParticipantCreate,
            exam_id: UUID = Path(...),
            db: Session = Depends(get_db),
            current_user: User = Depends(require_role('supervisor')),
        ):
            return self.service.add(db, exam_id, payload)

        @self.router.delete(
            "/{user_id}",
            summary="Видалити студента (завершити активну спробу, якщо йде тест)",
            status_code=status.HTTP_200_OK
        )
        async def remove_participant(
            user_id: UUID,
            exam_id: UUID = Path(...),
            db: Session = Depends(get_db),
            current_user: User = Depends(require_role('supervisor')),
        ):
            return self.service.remove(db, exam_id, user_id)

        @self.router.patch(
            "/{user_id}/attendance",
            response_model=ExamParticipantResponse,
            summary="Позначити студента як присутнього або відсутнього"
        )
        async def set_attendance(
            user_id: UUID,
            update: ExamParticipantAttendanceUpdate,
            exam_id: UUID = Path(...),
            db: Session = Depends(get_db),
            current_user: User = Depends(require_role('supervisor')),
        ):
            return self.service.set_attendance(db, exam_id, user_id, update)

    # Supervisor enforcement is provided by the `require_role('supervisor')` dependency